requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
authlib==1.2.1
itsdangerous==2.1.2
python-jose[cryptography]==3.3.0

# Environment variables
python-dotenv==1.0.0

# Image processing
Pillow==10.1.0

# HTTP requests (for Discord OAuth)
httpx==0.25.1

# Templates
jinja2==3.1.2

# Session management
starlette==0.27.0
